+++ NEWS.md
@@ -2,9 +2,19 @@
 
 ## New features
 
+* New parameter: `n_sgd_threads`, which controls the number of threads used
+in the stochastic gradient descent. By default this is now single-threaded
+and should result in reproducible results when using `set.seed`. To get back
+the old, less consistent, but faster settings, set `n_sgd_threads` 
+appropriately (e.g. `max(1, RcppParallel::defaultNumThreads() / 2)`).
 * API change for consistency with Python UMAP:
   * `alpha` is now `learning_rate`.
   * `gamma` is now `repulsion_strength`.
+
+## Bug fixes and minor improvements
+
+* Smooth knn/perplexity calibration results had a small dependency on the 
+number of threads used.
 
 # uwot 0.0.0.9007 (December 9 2018)